Move builtin methods to pathname.rbinc from lib/pathname.rb
