+++ inits.c
@@ -99,6 +99,7 @@ rb_call_builtin_inits(void)
     BUILTIN(ast);
     BUILTIN(trace_point);
     BUILTIN(pack);
+    BUILTIN(pathname);
     BUILTIN(warning);
     BUILTIN(array);
     BUILTIN(hash);
